whois-scanner/code/requirements.txt
PyYAML
python-whois
